Moved test_utilities.hpp definitions to separate .cpp file and updated CMakeLists.txt accordingly.
